requirements.txt
opencv-python
pytesseract
numpy
openai
robin-stocks
pyautogui
Pillow
pyotp
python-dotenv
cachetools